requirements.txt
PyPDF2
pypdf
python-dotenv
openai
chromadb
langchain
tiktoken
wikipedia
sentence_transformers
pdfminer
pdfminer-six
gensim
logger
seaborn
matplotlib
nltk
flask
requests
pandas